Melhorado a página de carrinho, criado a lógica para inserir os dados dos produtos na API, finalizado a função de remover um item do carrinho.

+++ src/services/api.ts
@@ -1,3 +1,4 @@
+import type { IPedido } from "../interface/pedidoInterface";
 import type { IProduto } from "../interface/produto-interface";
 
 
@@ -27,6 +28,20 @@ export async function adicionarProdutoAPI(produto: IProduto): Promise<IProduto> 
 }
 
 // ===== CARRINHO =====
+
+// Atualiza quantidade no carrinho
+export async function atualizarQuantidadeCarrinho(id: number, quantidade: number) {
+  const response = await fetch(`http://localhost:3000/carrinho/${id}`,{
+    method: "PATCH",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify({ quantidade })
+  });
+
+  return response.json();
+}
+
 export async function listarCarrinho(): Promise<ICarrinhoItem[]> {
     const response = await fetch(`${API_URL}/carrinho`);
     return await response.json();
@@ -49,18 +64,64 @@ export async function listarCarrinhoCompleto() {
   });
 }
 
-export async function adicionarAoCarrinho(item: ICarrinhoItem): Promise<ICarrinhoItem> {
-    const response = await fetch(`${API_URL}/carrinho`, {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(item)
-    });
+// ===== CHECKOUT ===== 
 
-    return await response.json();
+export async function criarPedido(pedido: IPedido){
+  const response = await fetch("http://localhost:3000/pedidos", {
+    method: "POST",
+    headers: {
+      "content-type": "application/json"
+    },
+    body: JSON.stringify(pedido)
+  });
+
+  if (!response.ok) {
+    throw new Error("Erro ao criar pedido");
+  }
+
+  return response.json();
+}
+
+export async function limpaCarrinho(){
+  const response = await fetch("http://localhost:3000/carrinho");
+  
+  if(!response.ok){
+    throw new Error("Erro ao buscar carrinho!");
+  }
+
+  const itens = await response.json();
+
+  await Promise.all(
+    itens.map((item: any) => {
+      fetch(`http://localhost:3000/carrinho/${item.id}`, {
+        method: "DELETE"
+      })
+    })
+  );
+}
+
+export async function adicionarAoCarrinho(
+  item: ICarrinhoItem
+): Promise<ICarrinhoItem> {
+  const response = await fetch(`${API_URL}/carrinho`, {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify(item)
+  });
+
+  if (!response.ok) {
+    throw new Error("Erro ao adicionar ao carrinho");
+  }
+
+  return response.json();
 }
 
 export async function removerDoCarrinho(id: number): Promise<void> {
-    await fetch(`${API_URL}/carrinho/${id}`, {
-        method: "DELETE"
-    });
+  const response = await fetch(`${API_URL}/carrinho/${id}`, {
+    method: "DELETE"
+  });
+
+  if (!response.ok) {
+    throw new Error("Erro ao remover item do carrinho");
+  }
 }
